Don't include datum just past requested date range
The data array returned by the fetcher used to also include the most
recent entry prior to the date range, since the check for the date was
made after the entry had been read into the array. Now, the first entry
past the requested date range is popped off before the data are
returned.