Make graph's horizontal margins smaller
The right margin was a bit too big; depending on the width of the y axis
labels.

At some point, I'll implement a smarter system for managing margin
widths relative to the sizes of the y axis labels. Problem is, I want
consistency across multiple graphs stacked vertically, so dynamic
margin sizes aren't a  good solution. I might implement label elision,
but the problem with that is you lose visual numeric precision. I might
add the option to specify margin sizes on the command-line, but that's
a bit cumbersome. We'll see.'